+++ src/cli/utils/sharepoint.go
@@ -54,22 +54,22 @@ func AddSharePointFilter(
 
 // IncludeSharePointRestoreDataSelectors builds the common data-selector
 // inclusions for SharePoint commands.
-func IncludeSharePointRestoreDataSelectors(
-	sel *selectors.SharePointRestore,
-	opts SharePointOpts,
-) {
+func IncludeSharePointRestoreDataSelectors(opts SharePointOpts) *selectors.SharePointRestore {
+	sites := opts.Sites
+
 	lp, li := len(opts.LibraryPaths), len(opts.LibraryItems)
 	ls, lwu := len(opts.Sites), len(opts.WebURLs)
 	slp, sli := len(opts.ListPaths), len(opts.ListItems)
 
 	if ls == 0 {
-		opts.Sites = selectors.Any()
+		sites = selectors.Any()
 	}
 
-	if lp+li+lwu+slp+sli == 0 {
-		sel.Include(sel.Sites(opts.Sites))
+	sel := selectors.NewSharePointRestore(sites)
 
-		return
+	if lp+li+lwu+slp+sli == 0 {
+		sel.Include(sel.Sites(sites))
+		return sel
 	}
 
 	if lp+li > 0 {
@@ -81,11 +81,11 @@ func IncludeSharePointRestoreDataSelectors(
 		containsFolders, prefixFolders := splitFoldersIntoContainsAndPrefix(opts.LibraryPaths)
 
 		if len(containsFolders) > 0 {
-			sel.Include(sel.LibraryItems(opts.Sites, containsFolders, opts.LibraryItems))
+			sel.Include(sel.LibraryItems(sites, containsFolders, opts.LibraryItems))
 		}
 
 		if len(prefixFolders) > 0 {
-			sel.Include(sel.LibraryItems(opts.Sites, prefixFolders, opts.LibraryItems, selectors.PrefixMatch()))
+			sel.Include(sel.LibraryItems(sites, prefixFolders, opts.LibraryItems, selectors.PrefixMatch()))
 		}
 	}
 
@@ -118,6 +118,8 @@ func IncludeSharePointRestoreDataSelectors(
 			sel.Include(sel.WebURL(suffixURLs, selectors.SuffixMatch()))
 		}
 	}
+
+	return sel
 }
 
 // FilterSharePointRestoreInfoSelectors builds the common info-selector filters.