+++ CHANGELOG.md
@@ -5,13 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] (alpha)
+
+### Changed
+
+- The selectors Reduce() process will only include details that match the DiscreteOwner, if one is specified.
+- New selector constructors will automatically set the DiscreteOwner if given a single-item slice.
 
 ### Fixed
 
 - Fixed issue where repository connect progress bar was clobbering backup/restore operation output.
+- Fixed issue where a `backup create exchange` produced one backup record per data type.
 
-## [v0.0.4] (alpha)
+### Known Issues
+
+- `backup list` will not display a resource owner for backups created prior to this release.
+
+## [v0.0.4] (alpha) - 2022-12-23
 
 ### Added